+++ sample-project/www/js/index.js
@@ -39,16 +39,19 @@ var app = {
         document.getElementById("refreshFeatureFlagsBtn").addEventListener("click", refreshFeatureFlags);
         document.getElementById("subscribeToFeatureFlagsBtn").addEventListener("click", subscribeToFeatureFlags);
         document.getElementById("getFeatureFlagPropertyBtn").addEventListener("click", getFeatureFlagProperty);
+        document.getElementById("logFeatureFlagImpressionBtn").addEventListener("click", logFeatureFlagImpression);
         document.getElementById("changeUserBtn").addEventListener("click", changeUser);
         document.getElementById("setSdkAuthBtn").addEventListener("click", setSdkAuthenticationSignature);
         document.getElementById("logCustomEventBtn").addEventListener("click", logCustomEvent);
         document.getElementById("logPurchaseBtn").addEventListener("click", logPurchase);
         document.getElementById("setCustomUserAttributeBtn").addEventListener("click", setCustomUserAttribute);
+        document.getElementById("mergeCustomUserAttributeBtn").addEventListener("click", mergeCustomUserAttribute);
         document.getElementById("setUserPropertiesBtn").addEventListener("click", setUserProperties);
         document.getElementById("launchNewsFeedBtn").addEventListener("click", launchNewsFeed);
         document.getElementById("launchContentCardsBtn").addEventListener("click", launchContentCards);
         document.getElementById("unsetCustomUserAttributeBtn").addEventListener("click", unsetCustomUserAttribute);
         document.getElementById("setCustomUserAttributeArrayBtn").addEventListener("click", setCustomUserAttributeArray);
+        document.getElementById("setCustomUserAttributeObjectArrayBtn").addEventListener("click", setCustomUserAttributeObjectsArray);
         document.getElementById("incrementCustomUserAttributeBtn").addEventListener("click", incrementCustomUserAttribute);
         document.getElementById("addToCustomUserAttributeArrayBtn").addEventListener("click", addToCustomUserAttributeArray);
         document.getElementById("removeFromCustomUserAttributeArrayBtn").addEventListener("click", removeFromCustomUserAttributeArray);
@@ -166,6 +169,16 @@ async function getFeatureFlagProperty() {
     }
 }
 
+function logFeatureFlagImpression() {
+    const featureFlagId = document.getElementById("featureFlagInputId").value;
+    if (!featureFlagId) {
+        showTextBubble("Feature Flag ID not entered.");
+        return;
+    }
+    BrazePlugin.logFeatureFlagImpression(featureFlagId);
+    showTextBubble(`Impression logged for FF ${featureFlagId}`);
+}
+
 function logCustomEvent() {
     const customEventID = document.getElementById("logCustomEventId").value;
     BrazePlugin.logCustomEvent(customEventID);
@@ -231,12 +244,53 @@ function logPurchase() {
 function setCustomUserAttribute() {
     BrazePlugin.setCustomUserAttribute("cordovaCustomAttributeKey", "cordovaCustomAttributeValue");
     BrazePlugin.incrementCustomUserAttribute("cordovaIncrementCustomAttributeKey", 1);
+
+    BrazePlugin.setCustomUserAttribute("CordovaNCA", {
+        "array key": [1, "two", false],
+        "object key": {
+            "k1": "one",
+            "k2": 2,
+            "k3": false,
+        },
+        "deep key": {
+            "key": [1, "two", true]
+        }
+    });
+
     showTextBubble("Set Custom User Attribute");
+}
+
+function mergeCustomUserAttribute() {
+    BrazePlugin.setCustomUserAttribute("CordovaNCA", 
+    {
+        "mergedInt": 1,
+        "object key": {
+            "mergek1": "1",
+            "mergek2": 2,
+            "mergek3": false,
+        }
+    },
+    true);
+    showTextBubble("Merge Custom User Attribute");
 }
 
 function setCustomUserAttributeArray() {
     BrazePlugin.setCustomUserAttribute("cordovaAttributeArrayButton", ["a", "b"]);
-    showTextBubble("Set Custom User Attribute Array");
+    showTextBubble("Set Custom User Attribute Strings Array");
+}
+
+function setCustomUserAttributeObjectsArray() {
+    BrazePlugin.setCustomUserAttribute("cordovaAttributeArrayButton", [
+        { 
+          "location": "East Rutherford, New Jersey",
+          "nickname": "Giants"
+        },
+        { 
+          "location": "Arlington, Texas",
+          "nickname": "Cowboys"
+        }  
+    ]);
+    showTextBubble("Set Custom User Attribute Objects Array");
 }
 
 function incrementCustomUserAttribute() {